Fixed bugs and got end-to-end working. - C++:  - Removed handcoded, as it is now a e2e test.         - Fixed bugs in module and instance to enable end-to-end.         - Added Fib as an end-to-end example test, which works.
- JS:   - Fixed bug causing stub to call incorrect method.
        - NaClModules can take a srcPrefix, used when inferring src.

- Test: - Added end to end test karma config.

- Makefile:
        - Greatly simplified Makefiles (again :P).
        - Added e2e test target.
        - Uses newlib, Debug, and x86_32 by default for everything.

+++ test/e2e/Fib/Fib.h
@@ -2,8 +2,8 @@
 #define FIB_H_
 
 namespace Fib{
-int fib(int x);
-int countUp();
+long fib(long x);
+long countUp();
 
 class MyClass{
 public:
@@ -20,7 +20,7 @@ public:
 	}
 
 private:
-	int counter;
+	long counter;
 };
 
 }
